// Das15 homework/src/types.ts
export interface IFormInput {
    name: string;
    surname: string;
    age: number;
    salary: number;
  }



 export interface IFormInput {
    name: string;
    surname: string;
    age: number;
    salary: number;
  }


export interface IUser {
    id: number;
    name: string;
    surname: string;
    age: number;
    salary: number;
  }